requirements.txt
lxml
python-dateutil
six
simpleeval<0.9.12
numpy>=1.8.0; python_version>="2.7"